+++ bytebank/src/main/kotlin/br/com/alura/bytebank/main.kt
@@ -1,47 +1,5 @@
-import br.com.alura.bytebank.modelo.Cliente
-import br.com.alura.bytebank.modelo.ContaPoupanca
-import br.com.alura.bytebank.modelo.Endereco
+import br.com.alura.bytebank.teste.testaHOF
 
 fun main() {
-    testaRun()
-}
-
-fun testaRun() {
-    val taxaAnual = 0.05
-    val taxaMensal = taxaAnual / 12
-    println("Taxa Mensal: $taxaMensal")
-
-    val contaPoupanca = ContaPoupanca(Cliente(nome = "Alex", cpf = "111.111.111-11", senha = 1234), 1000)
-
-    contaPoupanca
-        .run {
-            deposita(1000.0)
-            saldo * taxaMensal
-        }.let { rendimentoMensal ->
-            println("Rendimento Mensal: $rendimentoMensal")
-        }
-
-    val rendimentoAnual = run {
-        var saldo = contaPoupanca.saldo
-        repeat(12) { indice ->
-            saldo += saldo * taxaMensal
-        }
-        saldo
-    }
-    println("Rendimento Anual: $rendimentoAnual")
-}
-
-fun testaWith() {
-    with(Endereco()) {
-        logradouro = "rua vergueiro"
-        numero = 3185
-        bairro = "Vila Mariana"
-        cidade = "São Paulo"
-        estado = "SP"
-        cep = "02310-063"
-        complemento = "Apartamento"
-        completo()
-    }.let { enderecoCompleto: String ->
-        println(enderecoCompleto)
-    }
+    testaHOF()
 }
