+++ bytebank/src/main/kotlin/br/com/alura/bytebank/modelo/SistemaInterno.kt
@@ -22,4 +22,17 @@ class SistemaInterno {
             }
         }
     }
+
+    fun entraReceiver(admin: Autenticavel, senha: Int, autenticado: SistemaInterno.() -> Unit = {}) {
+        if (admin.autentica(senha)){
+            println("Bem vindo ao ByteBank!")
+            autenticado()
+        }else {
+            println("Falha na autenticação!")
+        }
+    }
+
+    fun pagamento(){
+        println("Realizando pagamento!")
+    }
 }
