+++ bytebank/src/main/kotlin/br/com/alura/bytebank/teste/testaRun.kt
@@ -1,0 +1,29 @@
+package br.com.alura.bytebank.teste
+
+import br.com.alura.bytebank.modelo.Cliente
+import br.com.alura.bytebank.modelo.ContaPoupanca
+
+fun testaRun() {
+    val taxaAnual = 0.05
+    val taxaMensal = taxaAnual / 12
+    println("Taxa Mensal: $taxaMensal")
+
+    val contaPoupanca = ContaPoupanca(Cliente(nome = "Alex", cpf = "111.111.111-11", senha = 1234), 1000)
+
+    contaPoupanca
+        .run {
+            deposita(1000.0)
+            saldo * taxaMensal
+        }.let { rendimentoMensal ->
+            println("Rendimento Mensal: $rendimentoMensal")
+        }
+
+    val rendimentoAnual = run {
+        var saldo = contaPoupanca.saldo
+        repeat(12) { indice ->
+            saldo += saldo * taxaMensal
+        }
+        saldo
+    }
+    println("Rendimento Anual: $rendimentoAnual")
+}
