+++ bytebank/src/main/kotlin/br/com/alura/bytebank/teste/testaWith.kt
@@ -1,0 +1,18 @@
+package br.com.alura.bytebank.teste
+
+import br.com.alura.bytebank.modelo.Endereco
+
+fun testaWith() {
+    with(Endereco()) {
+        logradouro = "rua vergueiro"
+        numero = 3185
+        bairro = "Vila Mariana"
+        cidade = "São Paulo"
+        estado = "SP"
+        cep = "02310-063"
+        complemento = "Apartamento"
+        completo()
+    }.let { enderecoCompleto: String ->
+        println(enderecoCompleto)
+    }
+}
